Whenever data is transmitted from one location to any other, both between modules in a single system or among separate computers, ensuring t
Whenever data is transmitted from one location to any other, both
between modules in a single system or among separate computers, ensuring
that the facts arrives at its destination without errors is of critical
significance. As a set of information is being transmitted over a
conversation line, ambient digital noise may additionally purpose one or
greater bits inside the transmission to turn out to be corrupted.
Frequently, this corruption takes the shape of a bit flip, both from a
one to a zero, or vice versa. The opportunity of one such errors
happening for an single bit in a given transmission may be very small,
however as the scale of the transmission will increase, so does the
chance of a bit flip taking place someplace within the transmission.
Detecting and/or correcting those bit flips is the purpose of every
error detection and correction code.
We use a technique called Hamming code for Data encoding and Error
Correction which is described in detail in this section.